backend/pyproject.toml
[project]
name = "entropic-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pyzmq>=26.0",
    "numpy>=2.0",
    "av>=14.0",
    "Pillow>=11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
